move url-building logic to component. Adjust operators to use component
