+++ app/components/url-builder/template.hbs
@@ -1,0 +1,1 @@
+<div><a target="_blank" href="{{queryUrl}}">See the Transitland API request URL for these results.</a></div>
