+++ app/operators/template.hbs
@@ -35,7 +35,7 @@
       {{/if}}
     {{/if}}
   </div>
-  <div><a target="_blank" href="{{queryUrl}}">See the Transitland API request URL for these results.</a></div>
+  {{url-builder model=operators type="operators" queryParams=queryParams bbox=bbox onestop_id=onestop_id}}
 </div>
 <div class="col-xs-12 col-sm-12 col-md-9">
   <div id="map">
